+++ test/lib/convert-forms/handle-no-label-placeholders.spec.js
@@ -1,8 +1,22 @@
 const { expect } = require('chai');
+const sinon = require('sinon');
+const log = require('../../../src/lib/log');
 const { removeNoLabelNodes } = require('../../../src/lib/convert-forms/handle-no-label-placeholders');
 const { createXformDoc, createXformString, serializeToString } = require('../../fn/convert-forms.utils');
 
+const WARNING_TEXT = 'The "NO_LABEL/DELETE_THIS_LINE" value is deprecated and will be removed in a future version of ' +
+  'cht-conf. For groups, a label is not required. If the field should not be visible, use the "hidden" or ' +
+  '"calculate" type.';
+
 describe('Handle NO_LABEL placeholders', () => {
+  beforeEach(() => {
+    sinon.stub(log, 'warn');
+  });
+
+  afterEach(() => {
+    sinon.restore();
+  });
+
   it('removes labels and itext text nodes that are only NO_LABEL', () => {
     const doc = createXformDoc({
       itext: `
@@ -54,6 +68,7 @@ describe('Handle NO_LABEL placeholders', () => {
       `
     });
     expect(serializeToString(doc)).xml.to.equal(expectedDoc);
+    expect(log.warn).to.have.been.calledOnceWithExactly(WARNING_TEXT);
   });
 
   it('keeps label when text has other values, but removes NO_LABEL values', () => {
@@ -91,6 +106,7 @@ describe('Handle NO_LABEL placeholders', () => {
       `
     });
     expect(serializeToString(doc)).xml.to.equal(expectedDoc);
+    expect(log.warn).to.have.been.calledOnceWithExactly(WARNING_TEXT);
   });
 
   it('handles multiple translations with NO_LABEL by removing all and associated labels', () => {
@@ -124,5 +140,6 @@ describe('Handle NO_LABEL placeholders', () => {
       body: `<input ref="/data/d"/>`
     });
     expect(serializeToString(doc)).xml.to.equal(expectedDoc);
+    expect(log.warn).to.have.been.calledOnceWithExactly(WARNING_TEXT);
   });
 });